feat: move Field/* theme files to root, remove keys
