+++ react/src/theme/components/YesNo.ts
@@ -2,8 +2,6 @@ import { anatomy, getColor, SystemStyleFunction } from '@chakra-ui/theme-tools'
 
 import { ComponentMultiStyleConfig } from '~/theme/types'
 
-export const YESNO_THEME_KEY = 'YesNoField'
-
 const parts = anatomy('rating').parts('option', 'container', 'icon')
 
 const outlineOptionStyle: SystemStyleFunction = (props) => {
@@ -85,7 +83,7 @@ const outlineContainerStyle: SystemStyleFunction = (props) => {
   }
 }
 
-export const YesNoField: ComponentMultiStyleConfig<typeof parts> = {
+export const YesNo: ComponentMultiStyleConfig<typeof parts> = {
   parts: parts.keys,
   variants: {
     outline: (props) => {
