+++ react/src/theme/components/index.ts
@@ -1,6 +1,4 @@
-import { Attachment, ATTACHMENT_THEME_KEY } from './Field/Attachment'
-import { RATING_THEME_KEY, RatingField } from './Field/Rating'
-import { YESNO_THEME_KEY, YesNoField } from './Field/YesNo'
+import { Attachment } from './Attachment'
 import { Avatar } from './Avatar'
 import { Badge } from './Badge'
 import { Banner } from './Banner'
@@ -22,6 +20,7 @@ import { NumberInput } from './NumberInput'
 import { Pagination, PAGINATION_THEME_KEY } from './Pagination'
 import { PhoneNumberInput } from './PhoneNumberInput'
 import { Radio, RADIO_THEME_KEY } from './Radio'
+import { Rating } from './Rating'
 import { Searchbar, SEARCHBAR_THEME_KEY } from './Searchbar'
 import { SingleCountryPhoneNumberInput } from './SingleCountryPhoneNumberInput'
 import { Tabs } from './Tabs'
@@ -31,6 +30,7 @@ import { Tile } from './Tile'
 import { Toast } from './Toast'
 import { Toggle, TOGGLE_THEME_KEY } from './Toggle'
 import { Tooltip } from './Tooltip'
+import { YesNo } from './YesNo'
 
 export const components = {
   Avatar,
@@ -55,14 +55,14 @@ export const components = {
   Textarea,
   Tabs,
   Tag,
-  [ATTACHMENT_THEME_KEY]: Attachment,
   [PAGINATION_THEME_KEY]: Pagination,
+  Attachment,
   [CHECKBOX_THEME_KEY]: Checkbox,
   [RADIO_THEME_KEY]: Radio,
   [SEARCHBAR_THEME_KEY]: Searchbar,
   Tooltip,
-  [RATING_THEME_KEY]: RatingField,
-  [YESNO_THEME_KEY]: YesNoField,
+  Rating,
+  YesNo,
   [TOGGLE_THEME_KEY]: Toggle,
   Tile,
   Toast,
